Quad Search Drawing Strategy
Start with one tile the size of the output image.  Sample the SDF at the
center.  If the sdf distance <= 0, and the abs(sdf distance) is >= the distance
from the center of the tile to one of teh corners, then draw a circle covering
the tile.  Otherwise, split the tile into four quadrants and recurse until the
tile cannot be split further.

This has ok performance for a small test scene and deterministic results, but
the edges of shapes look kind of rough.