"Divide and Monty" drawing strategy.
The divide-and-monty algorithm is like the quad-search algorithm, but the min
tile size should be set to be a little coarse (eg 32), and when the tile cannot
be split any further we sample N random points and draw distance circles if the
sdf distance is <= 0.

Quality degrades nicely with the number of random samples per tile.  It might
make sense to scale the number of samples based on the sample distance at the
center.

This approach struggles with tight points, and produces non-deterministic
results, but the edge quality for the most part looks a lot better than
the quad-search.