Random Split Option + Remove Redundant Color Changes
The random split option results in more iterations and draws, but requires
fewer random samples to get decent results.  This doesn't make much of a
difference in runtime perf, but it may be useful for other implementations.

Deduplicating color changes makes a pretty big difference though, halving
the overall time to run.